# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Supernova API
  description: |
    REST API for the Supernova task management application.

    ## Overview
    Supernova is a task app where your lock screen wallpaper IS your task list,
    showing you the ONE thing to do right now.

    ## Authentication
    Most endpoints require JWT authentication via Bearer token in the Authorization header.
    - **Access Token**: 7-day expiry, used for API access
    - **Refresh Token**: 30-day expiry, used to obtain new access tokens
    - **Wallpaper Token**: 365-day expiry, limited scope for widget access

    ## Rate Limiting
    - Global: 100 req/min
    - Auth endpoints: 10 req/min
    - Wallpaper: 20 req/hour
    - Task creation: 30 req/min

  version: 1.0.0
  contact:
    name: Supernova Support
    url: https://github.com/supernova-app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.supernova.app
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/user:
    get:
      tags:
        - User
      summary: Get user profile
      description: Retrieve current user's profile and settings
      operationId: getUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - User
      summary: Update user settings
      description: Update user profile settings. Invalidates wallpaper cache.
      operationId: updateUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - User
      summary: Delete user account
      description: Delete user account and all associated data (GDPR compliance)
      operationId: deleteUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Account deleted successfully. All associated data has been removed.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/user/export:
    get:
      tags:
        - User
      summary: Export user data
      description: Export all user data in JSON format (GDPR compliance)
      operationId: exportUserData
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User data export
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserExport'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/user/stats/weekly:
    get:
      tags:
        - User
      summary: Get weekly statistics
      description: Get usage statistics for the last 8 weeks with trends
      operationId: getWeeklyStats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Weekly statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/user/stats/graduation:
    get:
      tags:
        - User
      summary: Get graduation metrics
      description: Get graduation metrics showing app usage reduction over 6 months
      operationId: getGraduationStats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Graduation metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraduationStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/user/done-for-today:
    post:
      tags:
        - User
      summary: Mark done for today
      description: Mark user as done for the day (shows celebration wallpaper)
      operationId: markDoneForToday
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Marked as done
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  doneForToday:
                    type: boolean
                    example: true
                  doneAt:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # TASK ENDPOINTS
  # ============================================
...
  /api/done-for-today:
    post:
      tags:
        - Utility
      summary: Mark done for today (legacy)
      description: Legacy endpoint for marking done for today
      operationId: markDoneForTodayLegacy
      security:
        - bearerAuth: []
        - {}
      responses:
        '200':
          description: Marked as done
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              message: Authentication required
              code: AUTHENTICATION_ERROR
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        theme:
          type: string
          enum: [cosmic, ocean, fantasy]
        resolution:
          type: string
        display_mode:
          type: string
          enum: [one_thing, all_tasks]
        timezone:
          type: string
        setup_complete:
          type: boolean
        done_for_today:
          type: boolean
        done_for_today_at:
          type: string
          format: date-time
          nullable: true
        wallpaper_token:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        theme:
          type: string
          enum: [cosmic, ocean, fantasy]
        resolution:
          type: string
          pattern: '^\d+x\d+$'
        display_mode:
          type: string
          enum: [one_thing, all_tasks]
        timezone:
          type: string
        setup_complete:
          type: boolean
        done_for_today:
          type: boolean
    UserExport:
      type: object
      properties:
        exportedAt:
          type: string
          format: date-time
        version:
          type: string
          example: "1.0"
        user:
          $ref: '#/components/schemas/User'
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        stats:
          type: array
          items:
            $ref: '#/components/schemas/UserStat'
    UserStat:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        week_start:
          type: string
          format: date
        app_opens:
          type: integer
        widget_interactions:
          type: integer
        tasks_created:
          type: integer
        tasks_completed:
          type: integer
        tasks_completed_via_widget:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WeeklyStats:
      type: object
      properties:
        weeks:
          type: array
          items:
            $ref: '#/components/schemas/UserStat'
        trends:
          type: object
          properties:
            appOpenTrend:
              type: number
              nullable: true
              description: Percentage change in app opens
            widgetCompletionRate:
              type: number
              description: Percentage of tasks completed via widget
    GraduationStats:
      type: object
      properties:
        months:
          type: array
          items:
            type: object
            properties:
              month:
                type: string
                format: date-time
              total_app_opens:
                type: integer
              total_widget_interactions:
                type: integer
              total_tasks_completed:
                type: integer
              total_tasks_completed_via_widget:
                type: integer
        graduation:
          type: object
          properties:
            score:
              type: integer
              description: Graduation score (0-100)
            message:
              type: string
            widgetCompletionRate:
              type: number

    # ----------------------------------------
    # Task Schemas
    # ----------------------------------------
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        due_date:
          type: string
          format: date
          nullable: true
        due_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          nullable: true
        estimate_minutes:
          type: integer
          nullable: true
        priority:
          type: integer
          minimum: 1
          maximum: 3
          description: "1=critical, 2=normal, 3=low"
        completed:
          type: boolean
        completed_at:
          type: string
          format: date-time
          nullable: true
        snoozed_until:
          type: string
          format: date-time
          nullable: true
        context_location:
          type: string
          nullable: true
        context_time:
          type: string
          enum: [morning, afternoon, evening]
          nullable: true
        energy_required:
          type: string
          enum: [low, medium, high]
          nullable: true
        times_rescheduled:
          type: integer
        original_due_date:
          type: string
          format: date
          nullable: true
        x:
          type: number
          nullable: true
          description: Spatial coordinate
        y:
          type: number
          nullable: true
          description: Spatial coordinate
        is_subtask:
          type: boolean
        is_recurring:
          type: boolean
        echo_interval:
          type: string
          nullable: true
          description: "Recurrence interval (1d, 1w, etc.)"
        archived:
          type: boolean
        archived_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        score:
          type: integer
          description: Calculated priority score (higher = more urgent)
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - AUTHENTICATION_ERROR
                - AUTHORIZATION_ERROR
                - NOT_FOUND
                - CONFLICT
                - RATE_LIMIT
                - DATABASE_ERROR
                - CACHE_ERROR
                - INTERNAL_ERROR
            timestamp:
              type: string
              format: date-time
            path:
              type: string
            method:
              type: string
            details:
              type: object
    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
              msg:
                type: string
              path:
                type: string
              location:
                type: string